fix: unify coaching score display to 0.1° precision
Finish migrating Home/History/Care/Coaching score displays from the
old whole-degree ScoreCalculator.displayScore to the 0.1°-precision
displayValue, so the session-complete summary no longer shows a
different number than the home/history tabs for the same session.

+++ SmileDay/Views/Coaching/CoachingTabView.swift
@@ -8,8 +8,8 @@ struct CoachingTabView: View {
     @State private var result: SessionResult?
 
     struct SessionResult {
-        let todayScore: Int
-        let yesterdayScore: Int?
+        let todayScore: Double
+        let yesterdayScore: Double?
     }
 
     var body: some View {
